+++ __tests__/components/QueryForm.test.js
@@ -1,37 +1,70 @@
-import { mount } from '@vue/test-utils'
+import { mount, createLocalVue } from '@vue/test-utils'
+import VueRouter from 'vue-router'
+import QueryForm from 'src/components/QueryForm.vue'
 
-import ConfirmButton from 'src/components/ConfirmButton.vue'
+const localVue = createLocalVue()
 
-describe('NoMatch.vue', () => {
-  const wrapper = mount(ConfirmButton, {
-    propsData: {
-      text: '测试HOC',
-      onConfirmText: '你确定要删除吗？？',
-    },
-    slots: {
-      default: '<div class="fake-msg"></div>',
-    },
+localVue.use(VueRouter)
+
+const router = new VueRouter()
+
+describe('QueryForm.vue', () => {
+  let wrapper
+  beforeEach(() => {
+    wrapper = mount(QueryForm, {
+      propsData: {
+        showForm: {
+          name: true,
+          sex: true,
+          date: true,
+          age: true,
+        },
+      },
+      mocks: {
+        $t: () => 'some specific text',
+      },
+      router,
+      localVue,
+    })
   })
-  const mockFn = jest.fn(() => {})
-  wrapper.vm.$on('onConfirm', mockFn)
 
   it('测试接收到了prop参数', () => {
-    expect(wrapper.props().text).toContain('测试HOC')
-    expect(wrapper.props().onConfirmText).toContain('你确定要删除吗？？')
-  })
-
-  it('测试slot', () => {
-    const list = wrapper.find(ConfirmButton)
-    expect(list.contains('div.fake-msg')).toBeTruthy()
+    expect(wrapper.props().showForm).toEqual({
+      name: true,
+      sex: true,
+      date: true,
+      age: true,
+    })
   })
 
   it('测试onClick，是否触发事件', () => {
+    const mockFn = jest.fn(() => {})
+    wrapper.setMethods({
+      submitForm: mockFn,
+    })
     wrapper.vm.visible2 = true
     expect(mockFn).not.toHaveBeenCalled()
     wrapper
       .findAll('el-button')
-      .at(1)
+      .at(0)
       .trigger('click')
     expect(mockFn).toHaveBeenCalled()
+  })
+
+  it('测试resetQueryForm', () => {
+    const mockFn2 = jest.fn(() => {})
+    wrapper.setMethods({
+      resetForm: mockFn2,
+    })
+    wrapper.vm.ruleForm = {
+      name: '111',
+      sex: '111',
+      date: '2019-01-01',
+      age: '18',
+    }
+    const button = wrapper.findAll('el-button').at(1)
+    expect(mockFn2).not.toHaveBeenCalled()
+    button.trigger('click')
+    expect(mockFn2).toHaveBeenCalled()
   })
 })