+++ __tests__/tool/timeFormat.test.js
@@ -3,7 +3,7 @@ import timeFormat from 'tool/timeFormat'
 describe('测试传入Date对象，是否返回字符串', () => {
   it('测试传入Date对象', () => {
     const date = new Date('2018-11-01 16:37:52')
-    expect(timeFormat(date)).toBe('2018-11-01 16:37:52')
+    expect(timeFormat(date)).toBe('2018-11-01')
   })
 
   it('测试传入非Date对象', () => {