+++ src/store/userStore.js
@@ -15,7 +15,7 @@ const actions = {
   fetchLists: (store, params = { pageNo: 1, pageSize: 10 }) => {
     params.pageNo = params.pageNo ? params.pageNo : 1
     params.pageSize = params.pageSize ? params.pageSize : 10
-    axios.get(api.list, { params }).then(data => {
+    axios.get(api.list, { query: params }).then(data => {
       store.commit('setLists', data)
     })
   },