+++ src/tool/axios.js
@@ -1,35 +1,17 @@
-import axios from 'axios'
-import Vue from 'vue'
-import NProgress from 'nprogress'
+import { Fxios } from 'fxios'
+import appConfig from 'src/config'
 import isObject from 'lodash/isObject'
-import timeFormat from 'tool/timeFormat'
 import { emitter as soundEmitter } from '../components/SoundEffect.vue'
 
-axios.interceptors.request.use(
-  config => {
-    NProgress.start()
-    config.url = `/api/${config.url}`
-    /**
-     * 判断是否有Date格式的入参，统一转成字符串格式
-     */
-    if (config.params) {
-      Object.keys(config.params).forEach(item => {
-        if (config.params[item] instanceof Date) {
-          config.params[item] = timeFormat(config.params[item])
-        }
-      })
-    }
-    return config
-  },
-  err => {
-    Vue.prototype.$notify.error({
-      title: '内部错误',
-      message: err,
-      duration: 3000,
-    })
-    soundEmitter.emit('failure')
-  },
-)
+export const config = {
+  credentials: 'include',
+  redirect: 'manual',
+  mode: 'cors',
+  cache: 'reload',
+  base: appConfig.baseURL,
+}
+
+const axios = new Fxios(config)
 
 const processList = data => {
   if (isObject(data) && 'entities' in data) {
@@ -45,46 +27,32 @@ const processList = data => {
   return data
 }
 
-axios.interceptors.response.use(
-  res => {
-    NProgress.done()
-    if (res.data.code === '200') {
-      /**
-       * 如果没有返回数据，说明是用户操作，弹出提示框，并且增加声音
-       */
-      if (!res.data.data) {
-        soundEmitter.emit('success')
-        Vue.prototype.$notify.success({
-          title: '操作成功',
-          message: res.data.message,
-          duration: 3000,
-        })
-      }
-      return processList(res.data.data)
+axios.interceptor.response.push((res, req) => {
+  if (!res.ok) {
+    const error = new Error(res.statusText)
+    error.response = res
+    throw error
+  }
+  return res.json().then(data => {
+    res.message = data.message
+    if (req.method.toUpperCase() !== 'GET') {
+      axios.emit('success', res, req)
     }
-    Vue.prototype.$notify.error({
-      title: '接口错误',
-      message: res.data.message,
-      duration: 3000,
-    })
-    soundEmitter.emit('failure')
-    return null
-  },
-  err => {
-    if (err.response.status === 504 || err.response.status === 404) {
-      // Message.error({ message: '服务器被吃了⊙﹏⊙∥' })
-    } else if (err.response.status === 403) {
-      // Message.error({ message: '权限不足,请联系管理员!' })
-    } else {
-      // Message.error({ message: '未知错误!' })
+    if (!data.data) {
+      soundEmitter.emit('success')
     }
-    Vue.prototype.$notify.error({
-      title: '服务器错误',
-      message: err.title || err.toString(),
-      duration: 3000,
-    })
-    soundEmitter.emit('failure')
-  },
-)
+    return processList(data.data)
+  })
+})
+
+const fxiosCatch = error => {
+  soundEmitter.emit('failure')
+  // 若emitter没有监听函数直接emit一个error，会抛出错误不执行下面的throw error
+  if (axios.listeners('error').length > 0) {
+    axios.emit('error', error)
+  }
+  throw error
+}
+axios.interceptor.catch.push(fxiosCatch)
 
 export default axios