+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+## 0.1.2 (2018-11-29)
+
+
+### Bug Fixes
+
+* **fix templates:** 修改templates模板 ([1c266ef](https://github.com/yshysh123/vue-vuex-scaffold/commit/1c266ef))
+* **fixed queryform:** 修改queryForm的bug 完成度90% 目前有点就是router本页面query加不上的问题 用了另外一种实现方式 ([8aaa67e](https://github.com/yshysh123/vue-vuex-scaffold/commit/8aaa67e))
+* **fixed queryform vuex bug and update nodemon:** 修改查询框vuex的bug  升级nodemon ([cfa1253](https://github.com/yshysh123/vue-vuex-scaffold/commit/cfa1253))
+* **reset echarts:** 使用vue-echarts  修改路由跳转函数 ([72ef938](https://github.com/yshysh123/vue-vuex-scaffold/commit/72ef938))
+
+
+### Features
+
+* router store server ([f1ae010](https://github.com/yshysh123/vue-vuex-scaffold/commit/f1ae010))
+* store\server ([2444b57](https://github.com/yshysh123/vue-vuex-scaffold/commit/2444b57))
+* **add @babel/plugin-proposal-optional-chaining:** 增加链式操作符babel ([128d306](https://github.com/yshysh123/vue-vuex-scaffold/commit/128d306))
+* 增加node子模块 ([69ef4ab](https://github.com/yshysh123/vue-vuex-scaffold/commit/69ef4ab))
+* **404page confirmbutton:** 增加404页面 高阶组件测试 ([0afff36](https://github.com/yshysh123/vue-vuex-scaffold/commit/0afff36))
+* **add BundleAnalyzerPlugin user.page:** 增加打包体积插件 优化page页面 ([77de915](https://github.com/yshysh123/vue-vuex-scaffold/commit/77de915))
+* **add changelog  add e2e:** 增加changelog  增加e2e测试 ([d8d34b1](https://github.com/yshysh123/vue-vuex-scaffold/commit/d8d34b1))
+* **add e2e nightwatch:** 增加e2e测试 nightwatch  功能还需完善。流程已ok ([415460e](https://github.com/yshysh123/vue-vuex-scaffold/commit/415460e))
+* **add echarts:** 增加echarts简单封装 ([1355734](https://github.com/yshysh123/vue-vuex-scaffold/commit/1355734))
+* **add koa server:** 增加koa server 不断完善开发koa功能 ([5710d94](https://github.com/yshysh123/vue-vuex-scaffold/commit/5710d94))
+* **add loading  fix router:** 增加全局滚动条 修改路由跳转bug ([e9a48b0](https://github.com/yshysh123/vue-vuex-scaffold/commit/e9a48b0))
+* **add queryForm:** queryForm 增加url地址带参数 ([0070fef](https://github.com/yshysh123/vue-vuex-scaffold/commit/0070fef))
+* **add Skeleton:** 增加骨架屏 ([565e17c](https://github.com/yshysh123/vue-vuex-scaffold/commit/565e17c))
+* **add template:** 增加node模板 持续更新ing。功能尚未完善 ([9e186ba](https://github.com/yshysh123/vue-vuex-scaffold/commit/9e186ba))
+* **add templates  use yarn run addPage:** 增加node脚本 使用yarn run addPage 和yarn run deletePage来创建 删除页面 ([9fb2af6](https://github.com/yshysh123/vue-vuex-scaffold/commit/9fb2af6))
+* **add travis:** 增加travis工具（测试） ([13a8343](https://github.com/yshysh123/vue-vuex-scaffold/commit/13a8343))
+* **header sider colortheme:** header sider 增加theme  使用方法见header ([b509e12](https://github.com/yshysh123/vue-vuex-scaffold/commit/b509e12))
+* **i18n transition:** 增加国际化 路由动画 ([eed0e11](https://github.com/yshysh123/vue-vuex-scaffold/commit/eed0e11))
+* **sound:** add sound submit ([ac0f712](https://github.com/yshysh123/vue-vuex-scaffold/commit/ac0f712))
+* **theme css:** add theme  css  userqueryform ([bbe57c3](https://github.com/yshysh123/vue-vuex-scaffold/commit/bbe57c3))
+
+
+
 ## 0.1.1 (2018-11-20)